Refactor lap times and update calculation to have most recent lap first
